+++ jdbc/src/main/java/nextstep/jdbc/JdbcCallback.java
@@ -1,0 +1,9 @@
+package nextstep.jdbc;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public interface JdbcCallback<T> {
+
+    T call(PreparedStatement preparedStatement) throws SQLException;
+}
